新增修改密码

+++ web/src/main/java/com/cycredit/app/controller/memo/MemoController.java
@@ -285,7 +285,7 @@ public class MemoController {
             @ApiImplicitParam(name = "token", paramType = "header", value = "token", required = false),
             @ApiImplicitParam(name = "uid", paramType = "header", value = "uid", required = false),
     })
-    public Object departmentAdd(@RequestBody String f , @ApiParam("传入ID更新不传入则保存") @RequestParam(required = false, value = "id") Long id, @RequestParam(required = false, value = "memoId") Long memoId, @RequestParam(required = false, value = "departmentCode") String departmentCode, @RequestParam(required = false, value = "measure") String measure, @RequestParam(required = false, value = "reason") String reason) {
+    public Object departmentAdd(@RequestBody String f, @ApiParam("传入ID更新不传入则保存") @RequestParam(required = false, value = "id") Long id, @RequestParam(required = false, value = "memoId") Long memoId, @RequestParam(required = false, value = "departmentCode") String departmentCode, @RequestParam(required = false, value = "measure") String measure, @RequestParam(required = false, value = "reason") String reason) {
         UniMemoDepartment uniMemoDepartment = new UniMemoDepartment();
         uniMemoDepartment.setDepartmentCode(departmentCode);
         uniMemoDepartment.setMeasure(measure);
@@ -309,7 +309,12 @@ public class MemoController {
             @ApiImplicitParam(name = "uid", paramType = "header", value = "uid", required = false),
     })
     public Object departmentDelete(@RequestParam(required = false, value = "id") Long id) {
+        UniMemoDepartment uniMemoDepartment = memoDepartmentService.findMemoDepartmentById(id);
+        if (uniMemoDepartment == null) {
+            return Response.fail("删除失败，不存在的关联措施");
+        }
         memoDepartmentService.deleteDepartment(id);
+        memoDepartmentService.updateUnimo(uniMemoDepartment.getMemoId());
         return Response.success("成功");
 
     }
